# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.IncidentPostmortem:
    properties:
      action_items:
        description: JSON array of action items
        type: string
      contributing_factors:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      created_by_name:
        type: string
      id:
        type: integer
      lessons_learned:
        type: string
      problem_id:
        type: integer
      reviewed_at:
        type: string
      reviewed_by:
        type: integer
      reviewed_by_name:
        type: string
      root_cause:
        type: string
      ticket_id:
        type: integer
      ticket_subject:
        description: Transient join fields
        type: string
      timeline:
        description: JSON array of timeline events
        type: string
      updated_at:
        type: string
    type: object
  models.Problem:
    properties:
      assigned_to:
        type: integer
      created_at:
        type: string
      created_by:
        type: integer
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      known_error:
        type: boolean
      permanent_solution:
        type: string
      priority:
        description: Low | Medium | High | Critical
        type: string
      related_asset_id:
        type: integer
      resolved_at:
        type: string
      root_cause:
        type: string
      status:
        description: Open | Under Investigation | Known Error | Resolved | Closed
        type: string
      title:
        type: string
      updated_at:
        type: string
      updated_by:
        type: integer
      workaround:
        type: string
    type: object
  models.ProblemDetail:
    properties:
      assigned_to:
        type: integer
      assignee_name:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      created_by_name:
        type: string
      description:
        type: string
      id:
        type: integer
      incident_count:
        type: integer
      known_error:
        type: boolean
      linked_incidents:
        items:
          $ref: '#/definitions/models.ProblemIncident'
        type: array
      permanent_solution:
        type: string
      postmortem:
        $ref: '#/definitions/models.IncidentPostmortem'
      priority:
        type: string
      related_asset_id:
        type: integer
      related_asset_name:
        type: string
      resolved_at:
        type: string
      root_cause:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      workaround:
        type: string
    type: object
  models.ProblemIncident:
    properties:
      id:
        type: integer
      linked_at:
        type: string
      linked_by:
        type: integer
      linked_by_name:
        type: string
      notes:
        type: string
      problem_id:
        type: integer
      ticket_id:
        type: integer
      ticket_priority:
        type: string
      ticket_status:
        type: string
      ticket_subject:
        description: Transient join fields
        type: string
    type: object
  models.ProblemInfo:
    properties:
      assigned_to:
        type: integer
      assignee_name:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      created_by_name:
        type: string
      description:
        type: string
      id:
        type: integer
      incident_count:
        type: integer
      known_error:
        type: boolean
      priority:
        type: string
      related_asset_id:
        type: integer
      related_asset_name:
        type: string
      resolved_at:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API implementing ISO/IEC 19770-10 Grade A++ RBAC model
  title: IT Asset & Compliance API
  version: "1.1"
paths:
...
  /escalation-rules/{id}:
    delete:
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Hapus aturan eskalasi
      tags:
      - Escalation
    put:
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Update aturan eskalasi
      tags:
      - Escalation
...
  /problems:
    get:
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter known errors only
        in: query
        name: known_error
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProblemInfo'
            type: array
      summary: List semua problems
      tags:
      - Problem Management
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Problem'
      summary: Buat problem baru
      tags:
      - Problem Management
  /problems/{id}:
    delete:
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Soft-delete problem
      tags:
      - Problem Management
    get:
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProblemDetail'
      summary: Detail problem beserta linked incidents
      tags:
      - Problem Management
    put:
      consumes:
      - application/json
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Update problem (status, RCA, workaround, dll)
      tags:
      - Problem Management
  /problems/{id}/incidents:
    post:
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Hubungkan ticket (incident) ke problem
      tags:
      - Problem Management
  /problems/{id}/incidents/{ticket_id}:
    delete:
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ticket ID
        in: path
        name: ticket_id
        required: true
        type: integer
      responses: {}
      summary: Putus relasi ticket dari problem
      tags:
      - Problem Management
...
  /tickets/{ticket_id}/postmortem:
    get:
      parameters:
      - description: Ticket ID
        in: path
        name: ticket_id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Ambil post-mortem berdasarkan ticket
      tags:
      - Post-Mortem
    post:
      consumes:
      - application/json
      parameters:
      - description: Ticket ID
        in: path
        name: ticket_id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Buat post-mortem untuk incident (ticket)
      tags:
      - Post-Mortem
  /tickets/{ticket_id}/postmortem/review:
    post:
      parameters:
      - description: Ticket ID
        in: path
        name: ticket_id
        required: true
        type: integer
      responses: {}
      summary: Tandai post-mortem sudah di-review
      tags:
      - Post-Mortem
...
swagger: "2.0"